+++ test/src/jvmMain/kotlin/jp/juggler/konaArchive/KonaArchiveTestUtilsPermissionJvm.kt
@@ -1,0 +1,8 @@
+package jp.juggler.konaArchive
+
+internal actual fun mkdirWithUserRwxPermission(path: String) {
+    val directory = java.io.File(path)
+    if (!directory.mkdir() && !directory.isDirectory) {
+        throw java.io.IOException("Unable to create directory: $path")
+    }
+}