+++ test/src/linuxMain/kotlin/jp/juggler/konaArchive/KonaArchiveTestUtilsLinux.kt
@@ -1,5 +1,11 @@
 package jp.juggler.konaArchive
 
+import platform.posix.EEXIST
+import platform.posix.errno
 import platform.posix.mkdir
 
-internal actual fun mkdirWithUserRwxPermission(path: String): Int = mkdir(path, 448U)
+internal actual fun mkdirWithUserRwxPermission(path: String) {
+    if (mkdir(path, 448U) != 0 && errno != EEXIST) {
+        error("Unable to create directory: $path (errno=$errno)")
+    }
+}